% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcspp.R
\name{pcspp}
\alias{pcspp}
\title{PCS recodées pour comparatisme public-privé}
\usage{
pcspp(
  data,
  profession,
  admin_ent,
  nomenclaturePCS = c("2003", "1982", "2020"),
  sexeEE = 2015:2018,
  sexe,
  sexe_wt,
  threshold = c(33.3, 66.6),
  gender_sign = TRUE,
  masculiniser = FALSE,
  ESE = FALSE
)
}
\arguments{
\item{data}{La base de données dans laquelle effectuer les recodages.}

\item{profession}{Le nom de la variable professions détaillées. Souvent \code{P} ou \code{PE}
dans les bases fournies par l'INSEE.}

\item{admin_ent}{Le nom d'une variable catégorielle dont la première modalité
correspond aux administrations publiques (pour les différencier du secteur marchand).
Attention à ne pas inclure les entreprises publiques parmi les administrations.}

\item{nomenclaturePCS}{Quelle version de la nomemclature PCS la variable professions
utilise-t-elle ? \code{"2003"} par default, \code{"1982"} possible, 2020 non encore disponible.}

\item{sexeEE}{Par default, si aucune variable \code{sexe} n'est précisée, le calcul du taux
de féminisation est réalisé à partir de tableaux issus des enquêtes Emploi empilées
sur 1982-2018. Il est alors possible de sélectionner une année parmi celles disponibles,
ou de faire la moyenne sur plusieurs années, en renseignant une suite de nombre entiers
(integer vector). Par défault \code{2015:2018} (2015 à 2018).}

\item{sexe}{Une seconde possibilité est de fournir le nom d'une variable sexe présente
dans la base de données pour calculer le taux de féminisation. Pour être détectées les
modalités de variables doivent contenir les chaînes de caractères "femme"/"homme",
"masculin"/"féminin" ou "man"/"woman".}

\item{sexe_wt}{Si une variable \code{sexe} est fournie, il est possible de fournir le nom
de la variable de pondération à utiliser.}

\item{threshold}{Les seuils à utiliser pour calculer le taux de féminisation, sous
la forme d'une vecteur numérique de taille 1 ou 2. Le premier élément est le seuil
en-dessous duquel les noms de professions sont conjugués au masculin (par default 33.3\%).
Le second élément est le seuil en-dessus duquel ils sont conjugués au féminin
(par default 66.6\%). Entre les deux seuils, la forme inclusive avec point médian est
utilisée. Pour empêcher ce comportement, il est possible de ne préciser qu'un seul seuil.}

\item{gender_sign}{Par default, ajoute les symboles male et femelle lorsque le nom
d'une profession ou d'une CSP est épicène. Mettre sur \code{FALSE} pour éviter ce
comportement.}

\item{masculiniser}{Mettre sur \code{TRUE} pour ne choisir que des noms de professions
masculins.}

\item{ESE}{Par defaut, pour la PCS 2003, les catégories spécifiques aux PCS-ESE
sont repérées. Mettre sur \code{FALSE} pour les supprimer.}
}
\value{
La base de donnée de départ sous forme de \code{tibble}, avec cinq nouvelles
variables correspondant aux différents niveaux des PCS publiques-privées :
\itemize{
\item \code{PPP1} : le niveau agrégé (en 7 modalités) distingue "Chefs d’entreprise", "Cadres",
"Cols blancs subalternes", "Professionnel·les reconnu·es", "Employées", "Ouvriers"
et "Agriculteurs et ouvriers agricoles". Les consultants libéraux, professions libérales,
artisans des services, petits commerçants, artisans ouvriers, ont été classés dans
les catégories desquelles ils sont les plus proches du point de vue du revenu moyen,
du niveau de diplôme et du contenu de l'activité. Pour les enlever de l'analyse il est
nécessaire de filter les non-salariés.
\item \code{FAPPP} : les 11 filières professionnelles utilisées pour comparer public et privé.
\item \code{PPP2} : le niveau le plus détaillé (45 modalités) auquel il est possible de
constituer des catégories à la fois suffisamment homogène et regroupant salariés du
public et  salariés du privé. Il suffit alors de croiser la variable \code{PPP2} avec une
variable employeur dans un tableau pour effectuer une comparaison. Croisée avec les
filières professionnelles \code{FAPPP}, elle fait apparaître leurs hiérarchies internes.
\item \code{PPP3} : ce niveau situé entre les CSP et les professions (171 modalités) fait
apparaitre des professions appartenant aux administrations publiques et d'autres
appartenant au secteur marchand.
\item \code{PPP4} : les professions détaillées recodées au niveau le plus fin (459 modalités),
séparées entre administrations publiques et secteur marchand (pas encore de
féminisation personnalisable à ce niveau : celle adoptée est basée sur les chiffres de
l'enquête Emploi 2014-2018).
}
}
\description{
Ces catégories socioprofessionnelles, recodées à partir de la
\href{https://www.nomenclature-pcs.fr/}{nomenclature des PCS} de l'INSEE,
visent à faciliter la comparaison entre public et privé.
Il s'agit de construire des familles de métier comparables entre administrations publiques
et entreprises : les professions classées « Administration publique » par la
nomenclature FAP sont ventilées dans les familles professionnelles du secteur marchand,
en s'inspirant des
\href{https://www.fonction-publique.gouv.fr/files/files/ArchivePortailFP/www.fonction-publique.gouv.fr/la-nomenclature-fapfp-de-familles-de-metiers-de-la-fonction-publique.html}{nomenclature FaPFP des familles de métiers de la fonction publique}
de la DGAFP. Nous distinguons ainsi
différentes filières, commerciale, financière, administrative, technique, santé, etc.,
au sein desquelles une hiérarchie de métiers peut être comparée entre administrations
et entreprises. Les professions organisées comme les médecins, les infirmières ou
les enseignantes sont distingués des cols blancs (de niveau cadre ou de
niveau subalerne), car leur proportion s'avère très différence dans les entreprises
et dans les administrations. Certaines recommandations du Conseil national de
l'information statistique sont prises en compte (Amossé, Chardon et Eidelman, 2019) :
des "professions intermédiaires" de la fonction publique, classées A et apparentées
aux cadres du point du vue du diplôme et/ou du salaire, passent notamment au niveau
de qualification supérieur (cadres de santé, cadres du social, cadres de l’armée).
Pour rendre visible la place des femmes dans la structure sociale,
la règle de féminisation des noms de professions suivante est adoptée par défaut :
masculin s’il y a plus de deux tiers d’hommes dans une profession ; féminin s’il y a
plus de deux tiers de femmes ; forme inclusive s’il y a entre 33\% et 66\% de femmes.
}
